Check assumptions of weak superiority

This function tests if the probabilities within the Assignment Error Matrix are in line with the assumption of weak superiority.

Usage

check_conformity_c(aem)

Arguments

aem

matrix of probabilities

Value

Returns the number of violations of the assumption of weak superiority. 0 if the assumptions are fulfilled.

References

Berding, Florian, and Pargmann, Julia (2022).Iota Reliability Concept of the Second Generation.Measures for Content Analysis Done by Humans or Artificial Intelligences. Berlin: Logos. https://doi.org/10.30819/5581