Computes Iota and its elements in version 2

Computes all elements of the Iota Reliability Concept

compute_iota2(
  data,
  random_starts = 3,
  max_iterations = 5000,
  cr_rel_change = 1e-12,
  trace = TRUE
)

Arguments

data

Data for which the elements should be estimated. Data must be an object of type data.frame or matrix with cases in the rows and raters in the columns.

random_starts

An integer for the number of random starts for the EM-algorithm.

max_iterations

An integer for the maximum number of iterations within the EM-algorithm.

cr_rel_change

Positive numeric value for defining the convergence of the EM-algorithm.

trace

Boolean. TRUE if information on the estimation progress should be printed to the console.

Value

Returns a List with the following three components: The first component estimates_categorical_level comprises all elements that describe the ratings on a categorical level. The elements are sub-divided into raw estimates and chance-corrected estimates.

raw_estimates

alpha_reliability: A vector containing the Alpha Reliabilities for each category. These values represent probabilities.

beta_reliability: A vector containing the Beta Reliabilities for each category. These values represent probabilities.

assignment_error_matrix: AEM-matrix containing the conditional probabilities for assigning a unit of category i to categories 1 to n.

elements_chance_corrected

alpha_reliability: A vector containing the chance corrected Alpha Reliabilities for each category.

beta_reliability: A vector containing the chance corrected Beta Reliabilities for each category.

iota: A vector containing the Iota values for each category.

The second component estimates_scale_level contains elements for describing the quality of the ratings on a scale level. It comprises the following elements:

avg_iota: Numeric value representing the mean of the Iota values.

min_Iota: Numeric value representing the minimum of the Iota values.

total_iota: Numeric value representing the weighted mean of the Iota values. Class sizes are used for weighting.

aem_index: Numeric value representing the deviation of the AEM from a matrix describing random assignments.

aem_index_weighted: Numeric value representing the deviation of the AEM from a matrix describing random assignments. Categorical sizes are used for weighting. Nominal scale level is assumed.

aem_index_weighted_ord: Numeric value representing the deviation of the AEM from a matrix describing random assignments. Categorical sizes are used for weighting. Ordinal scale level is assumed.

The third component information contains important information regarding the parameter estimation. It comprises the following elements:

log_likelihood: Log-likelihood of the used solution.

Convergence: If estimation converged 0 else 1.

categorical_sizes: Estimates categorical sizes. This is the estimated amount of the categories.

conformity: 0 if the solution is in line with assumptions of weak superiority. A number greater 0 indicates the number of violations of the assumption of weak superiority..

n_condition: TRUE if the number of observed patterns equals or ist greater than the degrees of freedom. FALSE if not.

s_condition: TRUE if the rank of the derivates equals the degrees of freedoms. FALSE if not.

deviation: Measure for describing the distance between the different solutions. 0 indicates that all solutions generated for the different start values are the same.

n_pattern: Numer of observed patterns.

n_pattern_unordered: Numer of observed patterns if the order of codings are not considered.

n_parameter: Numer independently estimates parameters.

random_starts: Numer of random starts for the EM-algorithm.

Note

This is an experimental version of an improved version of the iota concept. Function and results may change in the future. We plan to implement the final version at the beginng of 2023.