Parameter estimation via EM-Algorithm

Function written in C++ for estimating the parameters of the model via Expectation-Maximization(EM)-algorithm.

EM_algo_c(
  obs_pattern_shape,
  obs_pattern_frq,
  obs_internal_count,
  categorical_levels,
  random_starts,
  max_iterations,
  rel_convergence,
  trace
)

Arguments

obs_pattern_shape

Matrix containing the unique patterns found in the data. Ideally this matrix is generated by the function get_patterns.

obs_pattern_frq

Vector containing the frequencies of the patterns. Ideally it is generated by the the function get_patterns.

obs_internal_count

Matrix containing the relative frequencies of each category within each pattern. Ideally this matrix is generated by the function get_patterns.

categorical_levels

Vector containing all possible categories of the content analysis.

random_starts

Integer for determining how often the algorithm should restart with randomly chosen values for the aem matrix and the categorical sizes.

max_iterations

Integer for determining the maximum number of iterations for each random start.

rel_convergence

Double for determining the convergence criterion. The algorithm stops if the relative change is smaller than this criterion.

trace

TRUE for printing progress information on the console. FALSE if this information should not be printed.

Value

Function returns a List with the estimated parameter sets for every random start. Every parameter set contains the following components:

log_likelihood

Log-likelihood of the estimated solution.

aem

Estimated assignment-error-matrix (aem). The rows represent the true categories while the columns stand for the assigned categories. The cells describe the probability that a coding unit of category i is assigned to category j.

categorial_sizes

Vector of estimated sizes for each category.

convergence

If the algorithm converged within the iteration limit TRUE. FALSE in every other case.

iteration

Number of iterations when the algorithm was terminated.