Computes Iota and its elements in version 1

Computes all elements of the Iota Reliability Concept

compute_iota1(data)

Arguments

data

Data for which the elements should be estimated. Data must be an object of type data.frame or matrix with cases in the rows and raters in the columns.

Value

A list with the following components

alpha

A vector containing the chance-corrected alpha reliabilities for every category.

beta

A vector containing the chance-corrected beta reliabilities for every category.

iota

A vector containing the iota values for every category.

assignment_error_matrix

A matrix with the conditional probabilities for every category. The rows refer to the true categories and the columns refer to the assigned categories. The elements on the diagonal represent the alpha errors of that category. The other elements in each row represent the conditioned probabilities that a coding unit is wrongly assigned to another category.

average_iota

A numeric value ranging between 0 and 1 representing the average iota values on a categorical level. It describes the reliability of the whole scale.

References

Berding, F., Riebenbauer, E., Stuetz, S., Jahncke, H., Slopinski, A., & Rebmann, K. (2022, March 23). Performance and Configuration of Artificial Intelligence in Business Education Learning Analytics Applications. A Content Analysis-Based Approach. https://doi.org/10.31235/osf.io/trvcy