Estimating log-likelihood

Function written in C++ estimating the log likelihood of a given parameter set.

Usage

fct_log_likelihood_c(
  categorial_sizes,
  aem,
  obs_pattern_shape,
  obs_pattern_frq,
  categorical_levels
)

Arguments

categorial_sizes

Vector containing the sizes of the different categories. That is amount of a category on all cases.

aem

Matrix in aem form. This matrix reports the true category in the rows and the assigned categories in the columns. The cells represent the probabilities that a coding unit of category i is assigned to category j.

obs_pattern_shape

Matrix containing the unique patterns found in the data. Ideally this matrix is generated by the function get_patterns().

obs_pattern_frq

Vector containing the frequencies of the patterns. Ideally it is generated by the the function get_patterns().

categorical_levels

Vector containing all possible categories of the content analysis.

Value

Returns the log likelihood as a single numeric value.

References

Berding, Florian, and Pargmann, Julia (2022).Iota Reliability Concept of the Second Generation.Measures for Content Analysis Done by Humans or Artificial Intelligences. Berlin: Logos. https://doi.org/10.30819/5581